WARNING: interim commit - WILL NOT COMPILE
This commit extracts part of the functionality of the LwjglRenderingProcess class into a new PostProcessor class. However, only methods and useful variables have been moved to the new class. The next commit will take care of patching things up so that the code can be successfully compiled.
