+++ engine/src/main/java/org/terasology/rendering/opengl/LwjglRenderingProcess.java
@@ -36,8 +36,8 @@ import org.terasology.registry.CoreRegistry;
 import org.terasology.rendering.assets.material.Material;
 import org.terasology.rendering.backdrop.BackdropProvider;
 import org.terasology.rendering.oculusVr.OculusVrHelper;
-import org.terasology.rendering.opengl.FBO.Dimensions;
 import org.terasology.rendering.world.WorldRenderer.WorldRenderingStage;
+import org.terasology.rendering.opengl.FBO.Dimensions;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
@@ -89,31 +89,6 @@ public class LwjglRenderingProcess {
 
     private static final Logger logger = LoggerFactory.getLogger(LwjglRenderingProcess.class);
 
-    /* PROPERTIES */
-    @EditorRange(min = 0.0f, max = 10.0f)
-    private float hdrExposureDefault = 2.5f;
-    @EditorRange(min = 0.0f, max = 10.0f)
-    private float hdrMaxExposure = 8.0f;
-    @EditorRange(min = 0.0f, max = 10.0f)
-    private float hdrMaxExposureNight = 8.0f;
-    @EditorRange(min = 0.0f, max = 10.0f)
-    private float hdrMinExposure = 1.0f;
-    @EditorRange(min = 0.0f, max = 4.0f)
-    private float hdrTargetLuminance = 1.0f;
-    @EditorRange(min = 0.0f, max = 0.5f)
-    private float hdrExposureAdjustmentSpeed = 0.05f;
-
-    @EditorRange(min = 0.0f, max = 5.0f)
-    private float bloomHighPassThreshold = 0.75f;
-    @EditorRange(min = 0.0f, max = 32.0f)
-    private float bloomBlurRadius = 12.0f;
-
-    @EditorRange(min = 0.0f, max = 16.0f)
-    private float overallBlurRadiusFactor = 0.8f;
-
-    /* HDR */
-    private float currentExposure = 2.0f;
-    private float currentSceneLuminance = 1.0f;
     private PBO readBackPBOFront;
     private PBO readBackPBOBack;
     private PBO readBackPBOCurrent;
@@ -138,7 +113,6 @@ public class LwjglRenderingProcess {
 
     /* VARIOUS */
     private boolean takeScreenshot;
-    private int displayListQuad = -1;
 
     // Note: this assumes that the settings in the configs might change at runtime,
     // but the config objects will not. At some point this might change, i.e. implementing presets.
@@ -260,625 +234,6 @@ public class LwjglRenderingProcess {
         }
     }
 
-    public boolean attachDepthBufferToFbo(String sourceFboName, String targetFboName) {
-        FBO source = getFBO(sourceFboName);
-        FBO target = getFBO(targetFboName);
-
-        if (source == null || target == null) {
-            return false;
-        }
-
-        glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, target.fboId);
-
-        glFramebufferRenderbufferEXT(GL_FRAMEBUFFER_EXT, GL_DEPTH_ATTACHMENT_EXT, GL_RENDERBUFFER_EXT, source.depthStencilRboId);
-        glFramebufferTexture2DEXT(GL_FRAMEBUFFER_EXT, GL_DEPTH_ATTACHMENT_EXT, GL11.GL_TEXTURE_2D, source.depthStencilTextureId, 0);
-
-        glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, 0);
-
-        return true;
-    }
-
-    private void updateExposure() {
-        if (renderingConfig.isEyeAdaptation()) {
-            FBO scene = getFBO("scene1");
-
-            if (scene == null) {
-                return;
-            }
-
-            readBackPBOCurrent.copyFromFBO(scene.fboId, 1, 1, GL12.GL_BGRA, GL11.GL_UNSIGNED_BYTE);
-
-            if (readBackPBOCurrent == readBackPBOFront) {
-                readBackPBOCurrent = readBackPBOBack;
-            } else {
-                readBackPBOCurrent = readBackPBOFront;
-            }
-
-            ByteBuffer pixels = readBackPBOCurrent.readBackPixels();
-
-            if (pixels.limit() < 3) {
-                logger.error("Failed to auto-update the exposure value.");
-                return;
-            }
-
-            currentSceneLuminance = 0.2126f * (pixels.get(2) & 0xFF) / 255.f + 0.7152f * (pixels.get(1) & 0xFF) / 255.f + 0.0722f * (pixels.get(0) & 0xFF) / 255.f;
-
-            float targetExposure = hdrMaxExposure;
-
-            if (currentSceneLuminance > 0) {
-                targetExposure = hdrTargetLuminance / currentSceneLuminance;
-            }
-
-            float maxExposure = hdrMaxExposure;
-
-            if (CoreRegistry.get(BackdropProvider.class).getDaylight() == 0.0) {
-                maxExposure = hdrMaxExposureNight;
-            }
-
-            if (targetExposure > maxExposure) {
-                targetExposure = maxExposure;
-            } else if (targetExposure < hdrMinExposure) {
-                targetExposure = hdrMinExposure;
-            }
-
-            currentExposure = (float) TeraMath.lerp(currentExposure, targetExposure, hdrExposureAdjustmentSpeed);
-
-        } else {
-            if (CoreRegistry.get(BackdropProvider.class).getDaylight() == 0.0) {
-                currentExposure = hdrMaxExposureNight;
-            } else {
-                currentExposure = hdrExposureDefault;
-            }
-        }
-    }
-
-    public void renderPreCombinedScene() {
-        createOrUpdateFullscreenFbos();
-
-        if (renderingConfig.isOutline()) {
-            generateSobel();
-        }
-
-        if (renderingConfig.isSsao()) {
-            generateSSAO();
-            generateBlurredSSAO();
-        }
-
-        generateCombinedScene();
-    }
-
-    public void renderPost(WorldRenderingStage worldRenderingStage) {
-        if (renderingConfig.isLightShafts()) {
-            PerformanceMonitor.startActivity("Rendering light shafts");
-            generateLightShafts();
-            PerformanceMonitor.endActivity();
-        }
-
-        PerformanceMonitor.startActivity("Pre-post processing");
-        generatePrePost();
-        PerformanceMonitor.endActivity();
-
-        if (renderingConfig.isEyeAdaptation()) {
-            PerformanceMonitor.startActivity("Rendering eye adaption");
-            generateDownsampledScene();
-            PerformanceMonitor.endActivity();
-        }
-
-        PerformanceMonitor.startActivity("Updating exposure");
-        updateExposure();
-        PerformanceMonitor.endActivity();
-
-        PerformanceMonitor.startActivity("Tone mapping");
-        generateToneMappedScene();
-        PerformanceMonitor.endActivity();
-
-        if (renderingConfig.isBloom()) {
-            PerformanceMonitor.startActivity("Applying bloom");
-            generateHighPass();
-            for (int i = 0; i < 3; i++) {
-                generateBloom(i);
-            }
-            PerformanceMonitor.endActivity();
-        }
-
-        PerformanceMonitor.startActivity("Applying blur");
-        for (int i = 0; i < 2; i++) {
-            if (renderingConfig.getBlurIntensity() != 0) {
-                generateBlur(i);
-            }
-        }
-        PerformanceMonitor.endActivity();
-
-        PerformanceMonitor.startActivity("Rendering final scene");
-        if (worldRenderingStage == WorldRenderingStage.LEFT_EYE
-                || worldRenderingStage == WorldRenderingStage.RIGHT_EYE
-                || (worldRenderingStage == WorldRenderingStage.MONO && takeScreenshot)) {
-
-            renderFinalSceneToRT(worldRenderingStage);
-
-            if (takeScreenshot) {
-                saveScreenshot();
-            }
-        }
-
-        if (worldRenderingStage == WorldRenderingStage.MONO
-                || worldRenderingStage == WorldRenderingStage.RIGHT_EYE) {
-            renderFinalScene();
-        }
-        PerformanceMonitor.endActivity();
-    }
-
-    private void renderFinalSceneToRT(WorldRenderingStage renderingStage) {
-        Material material;
-
-        if (renderingDebugConfig.isEnabled()) {
-            material = Assets.getMaterial("engine:prog.debug").get();
-        } else {
-            material = Assets.getMaterial("engine:prog.post").get();
-        }
-
-        material.enable();
-
-        bindFbo("sceneFinal");
-
-        if (renderingStage == WorldRenderingStage.MONO || renderingStage == WorldRenderingStage.LEFT_EYE) {
-            glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-        }
-
-        switch (renderingStage) {
-            case MONO:
-                renderFullscreenQuad(0, 0, fullScale.width(), fullScale.height());
-                break;
-            case LEFT_EYE:
-                renderFullscreenQuad(0, 0, fullScale.width() / 2, fullScale.height());
-                break;
-            case RIGHT_EYE:
-                renderFullscreenQuad(fullScale.width() / 2, 0, fullScale.width() / 2, fullScale.height());
-                break;
-        }
-
-        unbindFbo("sceneFinal");
-    }
-
-    private void updateOcShaderParametersForVP(Material program, int vpX, int vpY, int vpWidth, int vpHeight, WorldRenderingStage renderingStage) {
-        float w = (float) vpWidth / fullScale.width();
-        float h = (float) vpHeight / fullScale.height();
-        float x = (float) vpX / fullScale.width();
-        float y = (float) vpY / fullScale.height();
-
-        float as = (float) vpWidth / vpHeight;
-
-        program.setFloat4("ocHmdWarpParam", OculusVrHelper.getDistortionParams()[0], OculusVrHelper.getDistortionParams()[1],
-                OculusVrHelper.getDistortionParams()[2], OculusVrHelper.getDistortionParams()[3], true);
-
-        float ocLensCenter = (renderingStage == WorldRenderingStage.RIGHT_EYE) ? -1.0f * OculusVrHelper.getLensViewportShift() : OculusVrHelper.getLensViewportShift();
-
-        program.setFloat2("ocLensCenter", x + (w + ocLensCenter * 0.5f) * 0.5f, y + h * 0.5f, true);
-        program.setFloat2("ocScreenCenter", x + w * 0.5f, y + h * 0.5f, true);
-
-        float scaleFactor = 1.0f / OculusVrHelper.getScaleFactor();
-
-        program.setFloat2("ocScale", (w / 2) * scaleFactor, (h / 2) * scaleFactor * as, true);
-        program.setFloat2("ocScaleIn", (2 / w), (2 / h) / as, true);
-    }
-
-    private void renderFinalScene() {
-
-        Material material;
-
-        if (renderingConfig.isOculusVrSupport()) {
-            material = Assets.getMaterial("engine:prog.ocDistortion").get();
-            material.enable();
-
-            updateOcShaderParametersForVP(material, 0, 0, fullScale.width() / 2, fullScale.height(), WorldRenderingStage.LEFT_EYE);
-        } else {
-            if (renderingDebugConfig.isEnabled()) {
-                material = Assets.getMaterial("engine:prog.debug").get();
-            } else {
-                material = Assets.getMaterial("engine:prog.post").get();
-            }
-
-            material.enable();
-        }
-
-        renderFullscreenQuad(0, 0, org.lwjgl.opengl.Display.getWidth(), org.lwjgl.opengl.Display.getHeight());
-
-        if (renderingConfig.isOculusVrSupport()) {
-            updateOcShaderParametersForVP(material, fullScale.width() / 2, 0, fullScale.width() / 2, fullScale.height(), WorldRenderingStage.RIGHT_EYE);
-
-            renderFullscreenQuad(0, 0, org.lwjgl.opengl.Display.getWidth(), Display.getHeight());
-        }
-    }
-
-    private void generateCombinedScene() {
-        Assets.getMaterial("engine:prog.combine").get().enable();
-
-        bindFbo("sceneOpaquePingPong");
-
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        unbindFbo("sceneOpaquePingPong");
-
-        flipPingPongFbo("sceneOpaque");
-        attachDepthBufferToFbo("sceneOpaque", "sceneReflectiveRefractive");
-    }
-
-    public void applyLightBufferPass(String target) {
-        Material program = Assets.getMaterial("engine:prog.lightBufferPass").get();
-        program.enable();
-
-        FBO targetFbo = getFBO(target);
-
-        int texId = 0;
-        if (targetFbo != null) {
-            GL13.glActiveTexture(GL13.GL_TEXTURE0 + texId);
-            targetFbo.bindTexture();
-            program.setInt("texSceneOpaque", texId++);
-
-            GL13.glActiveTexture(GL13.GL_TEXTURE0 + texId);
-            targetFbo.bindDepthTexture();
-            program.setInt("texSceneOpaqueDepth", texId++);
-
-            GL13.glActiveTexture(GL13.GL_TEXTURE0 + texId);
-            targetFbo.bindNormalsTexture();
-            program.setInt("texSceneOpaqueNormals", texId++);
-
-            GL13.glActiveTexture(GL13.GL_TEXTURE0 + texId);
-            targetFbo.bindLightBufferTexture();
-            program.setInt("texSceneOpaqueLightBuffer", texId++, true);
-        }
-
-        FBO targetPingPong = getFBO(target + "PingPong");
-        targetPingPong.bind();
-        graphicState.setRenderBufferMask(targetPingPong, true, true, true);
-
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        unbindFbo(target + "PingPong");
-
-        flipPingPongFbo(target);
-
-        if (target.equals("sceneOpaque")) {
-            attachDepthBufferToFbo("sceneOpaque", "sceneReflectiveRefractive");
-        }
-    }
-
-    public void generateSkyBand(int id) {
-        FBO skyBand = getFBO("sceneSkyBand" + id);
-
-        if (skyBand == null) {
-            return;
-        }
-
-        skyBand.bind();
-        graphicState.setRenderBufferMask(skyBand, true, false, false);
-
-        Material material = Assets.getMaterial("engine:prog.blur").get();
-
-        material.enable();
-        material.setFloat("radius", 8.0f, true);
-        material.setFloat2("texelSize", 1.0f / skyBand.width(), 1.0f / skyBand.height(), true);
-
-        if (id == 0) {
-            bindFboTexture("sceneOpaque");
-        } else {
-            bindFboTexture("sceneSkyBand" + (id - 1));
-        }
-
-        setViewportTo(skyBand.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        skyBand.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generateToneMappedScene() {
-        Assets.getMaterial("engine:prog.hdr").get().enable();
-
-        bindFbo("sceneToneMapped");
-
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        unbindFbo("sceneToneMapped");
-    }
-
-    private void generateLightShafts() {
-        Assets.getMaterial("engine:prog.lightshaft").get().enable();
-
-        FBO lightshaft = getFBO("lightShafts");
-
-        if (lightshaft == null) {
-            return;
-        }
-
-        lightshaft.bind();
-
-        setViewportTo(lightshaft.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        lightshaft.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generateSSAO() {
-        Material ssaoShader = Assets.getMaterial("engine:prog.ssao").get();
-        ssaoShader.enable();
-
-        FBO ssao = getFBO("ssao");
-
-        if (ssao == null) {
-            return;
-        }
-
-        ssaoShader.setFloat2("texelSize", 1.0f / ssao.width(), 1.0f / ssao.height(), true);
-        ssaoShader.setFloat2("noiseTexelSize", 1.0f / 4.0f, 1.0f / 4.0f, true);
-
-        ssao.bind();
-
-        setViewportTo(ssao.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        ssao.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generateSobel() {
-        Assets.getMaterial("engine:prog.sobel").get().enable();
-
-        FBO sobel = getFBO("sobel");
-
-        if (sobel == null) {
-            return;
-        }
-
-        sobel.bind();
-
-        setViewportTo(sobel.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        sobel.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generateBlurredSSAO() {
-        Material shader = Assets.getMaterial("engine:prog.ssaoBlur").get();
-        shader.enable();
-
-        FBO ssao = getFBO("ssaoBlurred");
-
-        if (ssao == null) {
-            return;
-        }
-
-        shader.setFloat2("texelSize", 1.0f / ssao.width(), 1.0f / ssao.height(), true);
-        ssao.bind();
-
-        setViewportTo(ssao.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-        bindFboTexture("ssao");
-
-        renderFullscreenQuad();
-
-        ssao.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generatePrePost() {
-        Assets.getMaterial("engine:prog.prePost").get().enable();
-
-        bindFbo("scenePrePost");
-
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        unbindFbo("scenePrePost");
-    }
-
-    private void generateHighPass() {
-        Material program = Assets.getMaterial("engine:prog.highp").get();
-        program.setFloat("highPassThreshold", bloomHighPassThreshold, true);
-        program.enable();
-
-        FBO highPass = getFBO("sceneHighPass");
-
-        if (highPass == null) {
-            return;
-        }
-
-        highPass.bind();
-
-        FBO sceneOpaque = getFBO("sceneOpaque");
-
-        int texId = 0;
-        GL13.glActiveTexture(GL13.GL_TEXTURE0 + texId);
-        sceneOpaque.bindTexture();
-        program.setInt("tex", texId++);
-
-//        GL13.glActiveTexture(GL13.GL_TEXTURE0 + texId);
-//        sceneOpaque.bindDepthTexture();
-//        program.setInt("texDepth", texId++);
-
-        setViewportTo(highPass.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        renderFullscreenQuad();
-
-        highPass.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generateBlur(int id) {
-        Material material = Assets.getMaterial("engine:prog.blur").get();
-        material.enable();
-
-        material.setFloat("radius", overallBlurRadiusFactor * renderingConfig.getBlurRadius(), true);
-
-        FBO blur = getFBO("sceneBlur" + id);
-
-        if (blur == null) {
-            return;
-        }
-
-        material.setFloat2("texelSize", 1.0f / blur.width(), 1.0f / blur.height(), true);
-
-        blur.bind();
-
-        setViewportTo(blur.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        if (id == 0) {
-            bindFboTexture("sceneToneMapped");
-        } else {
-            bindFboTexture("sceneBlur" + (id - 1));
-        }
-
-        renderFullscreenQuad();
-
-        blur.unbind();
-
-        setViewportToFullSize();
-    }
-
-    private void generateBloom(int id) {
-        Material shader = Assets.getMaterial("engine:prog.blur").get();
-
-        shader.enable();
-        shader.setFloat("radius", bloomBlurRadius, true);
-
-        FBO bloom = getFBO("sceneBloom" + id);
-
-        if (bloom == null) {
-            return;
-        }
-
-        shader.setFloat2("texelSize", 1.0f / bloom.width(), 1.0f / bloom.height(), true);
-
-        bloom.bind();
-
-        setViewportTo(bloom.dimensions());
-        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-        if (id == 0) {
-            getFBO("sceneHighPass").bindTexture();
-        } else {
-            getFBO("sceneBloom" + (id - 1)).bindTexture();
-        }
-
-        renderFullscreenQuad();
-
-        bloom.unbind();
-        setViewportToFullSize();
-    }
-
-    private void generateDownsampledScene() {
-        Material shader = Assets.getMaterial("engine:prog.down").get();
-        shader.enable();
-
-        for (int i = 4; i >= 0; i--) {
-            int sizePrev = TeraMath.pow(2, i + 1);
-
-            int size = TeraMath.pow(2, i);
-            shader.setFloat("size", size, true);
-
-            bindFbo("scene" + size);
-            glViewport(0, 0, size, size);
-
-            glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-            if (i == 4) {
-                bindFboTexture("scenePrePost");
-            } else {
-                bindFboTexture("scene" + sizePrev);
-            }
-
-            renderFullscreenQuad();
-
-            unbindFbo("scene" + size);
-        }
-
-        setViewportToFullSize();
-    }
-
-    public void renderFullscreenQuad() {
-        glMatrixMode(GL_MODELVIEW);
-        glPushMatrix();
-        glLoadIdentity();
-
-        glMatrixMode(GL_PROJECTION);
-        glPushMatrix();
-        glLoadIdentity();
-
-        renderQuad();
-
-        glPopMatrix();
-
-        glMatrixMode(GL_MODELVIEW);
-        glPopMatrix();
-    }
-
-    public void renderFullscreenQuad(int x, int y, int viewportWidth, int viewportHeight) {
-        glViewport(x, y, viewportWidth, viewportHeight);
-
-        glMatrixMode(GL_MODELVIEW);
-        glPushMatrix();
-        glLoadIdentity();
-
-        glMatrixMode(GL_PROJECTION);
-        glPushMatrix();
-        glLoadIdentity();
-
-        renderQuad();
-
-        glPopMatrix();
-
-        glMatrixMode(GL_MODELVIEW);
-        glPopMatrix();
-    }
-
-    private void renderQuad() {
-        if (displayListQuad == -1) {
-            displayListQuad = glGenLists(1);
-
-            glNewList(displayListQuad, GL11.GL_COMPILE);
-
-            glBegin(GL_QUADS);
-            glColor4f(1.0f, 1.0f, 1.0f, 1.0f);
-
-            glTexCoord2d(0.0, 0.0);
-            glVertex3i(-1, -1, -1);
-
-            glTexCoord2d(1.0, 0.0);
-            glVertex3i(1, -1, -1);
-
-            glTexCoord2d(1.0, 1.0);
-            glVertex3i(1, 1, -1);
-
-            glTexCoord2d(0.0, 1.0);
-            glVertex3i(-1, 1, -1);
-
-            glEnd();
-
-            glEndList();
-        }
-
-        glCallList(displayListQuad);
-    }
-
     public void takeScreenshot() {
         takeScreenshot = true;
 
@@ -941,10 +296,6 @@ public class LwjglRenderingProcess {
         overwriteRtHeight = 0;
 
         createOrUpdateFullscreenFbos();
-    }
-
-    public float getExposure() {
-        return currentExposure;
     }
 
     public FBO getFBO(String title) {
@@ -1041,15 +392,6 @@ public class LwjglRenderingProcess {
 
         fboLookup.put(title, fbo2);
         fboLookup.put(title + "PingPong", fbo1);
-    }
-
-
-    private void setViewportToFullSize() {
-        glViewport(0, 0, fullScale.width(), fullScale.height());
-    }
-
-    private void setViewportTo(Dimensions dimensions) {
-        glViewport(0, 0, dimensions.width(), dimensions.height());
     }
 
     /**
